+++ src/components/BarChart/BarChart.tsx
@@ -98,6 +98,7 @@ export function BarChart({
     showHorizontalLines: true,
     color: colorSky,
     horizontalOverflow: false,
+    horizontalMargin: 0,
     ...gridOptions,
   };
 
@@ -112,6 +113,7 @@ export function BarChart({
   const yAxisOptionsWithDefaults = {
     labelFormatter: (value: number) => value.toString(),
     labelColor: DEFAULT_GREY_LABEL,
+    backgroundColor: 'transparent',
     ...yAxisOptions,
   };