+++ src/components/LineChart/LineChart.tsx
@@ -93,6 +93,7 @@ export function LineChart({
   const yAxisOptionsWithDefaults = {
     labelFormatter: (value: number) => value.toString(),
     labelColor: DEFAULT_GREY_LABEL,
+    backgroundColor: 'transparent',
     ...yAxisOptions,
   };
 
@@ -101,6 +102,7 @@ export function LineChart({
     showHorizontalLines: true,
     color: colorSky,
     horizontalOverflow: false,
+    horizontalMargin: 0,
     ...gridOptions,
   };