+++ src/components/LineChart/hooks/use-y-scale.ts
@@ -3,7 +3,7 @@ import {scaleLinear} from 'd3-scale';
 
 import {getTextWidth} from '../../../utilities';
 import {yAxisMinMax} from '../utilities';
-import {MIN_Y_LABEL_SPACE, SPACING} from '../constants';
+import {MIN_Y_LABEL_SPACE} from '../constants';
 import {Series} from '../types';
 import {NumberLabelFormatter} from '../../../types';
 
@@ -37,13 +37,11 @@ export function useYScale({
       yOffset: yScale(value),
     }));
 
-    const maxTickWidth = Math.max(
+    const axisMargin = Math.max(
       ...ticks.map(({formattedValue}) =>
         getTextWidth({fontSize, text: formattedValue}),
       ),
     );
-
-    const axisMargin = maxTickWidth + SPACING;
 
     return {yScale, ticks, axisMargin};
   }, [series, drawableHeight, formatYAxisLabel, fontSize]);