+++ src/components/MultiSeriesBarChart/types.ts
@@ -52,6 +52,7 @@ export interface GridOptions {
   showHorizontalLines: boolean;
   color: string;
   horizontalOverflow: boolean;
+  horizontalMargin: number;
 }
 
 export interface XAxisOptions {
@@ -64,4 +65,5 @@ export interface XAxisOptions {
 export interface YAxisOptions {
   labelFormatter: NumberLabelFormatter;
   labelColor: string;
+  backgroundColor: string;
 }